species_evaluation_web/src/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic

numpy
pandas
scikit-learn
joblib

sqlalchemy
pymysql
python-dotenv
